chore: move cli.js into src/fe/cli.ts
This PR also adds a script to launch to the cli:

```shell
npm run madwizard <input file or uri>
```
