+++ .gitignore
@@ -3,3 +3,4 @@ dist
 .log
 node_modules
 *.tgz
+.parcel-cache
